% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_imgcoast.R
\name{load_imgcoast}
\alias{load_imgcoast}
\title{load_imgcoast
Read and performs pre-processing of image}
\usage{
load_imgcoast(x, target_ex, resolution = 0.5, crs = "+init=epsg:32632")
}
\arguments{
\item{target_ex}{Extent* raster object [ xmin ymin xmax ymax] .}

\item{resolution}{numeric Final resolution of raster object. default is 0.5 meter}

\item{crs}{character proj4 definition.}

\item{r}{character Path of coastal image to be processed.}
}
\description{
load_imgcoast
Read and performs pre-processing of image
}
